08-reasoning/pyproject.toml
[project]
name = "agenttools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biomcp-python>=0.1.2",
    "loguru>=0.7.3",
    "mcp[cli]>=1.6.0",
    "openai-agents>=0.0.9",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
